Create montecarloscc.jl
This function runs a Monte Carlo Simulation, using the distribution function from Hope (2009), described in mcs.jl, and includes a post trial function that calculates de SCC using the function in compute_scc.jl, It inputs the eta and prtp every time. Because of reasons that I do not know, if I'm honest, if the values are not inputted, the values of the SCC are not correctly computed, even though the function has a try statement which should include the value generated in every run.